Stabilize block breaking, block placing and item use
Promote block interaction handlers from experimental to the stable protocol: stable USE_ITEM_ON (block placing with StartItemUseOn/StopItemUseOn and creative-aware prediction) and USE_ITEM (with StartUsingItem tracking), drop and release via inventory transactions, real facing directions for block aborts, and clientbound inventory transaction application. Move the inventory transaction machinery to stable packages and remove the experimental packet factory.
