+++ README.md
@@ -39,9 +39,9 @@ Here is an overview of the current and planned features in ViaBedrock.
 - [ ] Client-Authoritative Inventory
 - [ ] Server-Authoritative Inventory
 - [ ] Item data
-- [ ] Block breaking
-- [ ] Block placing
-- [ ] Item use
+- [x] Block breaking
+- [x] Block placing
+- [x] Item use
 - [x] Respawning and dimension switching
 - [x] Form GUIs
 - [x] Scoreboard
@@ -56,8 +56,6 @@ Here is an overview of the current and planned features in ViaBedrock.
 ### Experimental
 Some features are experimental, which means they are almost certainly not fully stable/tested and may cause unexpected issues. To enable those features set `enable-experimental-features` to `true` in the config file.
 
-* Block placing
-* Item use
 * Some item data
 
 ## Optional clientside mods
