+++ src/main/java/net/raphimc/viabedrock/experimental/ExperimentalFeatures.java
@@ -18,343 +18,38 @@
 package net.raphimc.viabedrock.experimental;
 
 import com.viaversion.viaversion.api.connection.UserConnection;
-import com.viaversion.viaversion.api.minecraft.BlockFace;
 import com.viaversion.viaversion.api.minecraft.BlockPosition;
 import com.viaversion.viaversion.api.protocol.packet.PacketWrapper;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.libs.fastutil.longs.LongArrayList;
 import com.viaversion.viaversion.libs.fastutil.longs.LongList;
 import com.viaversion.viaversion.protocols.v1_21_11to26_1.packet.ClientboundPackets26_1;
-import com.viaversion.viaversion.protocols.v1_21_11to26_1.packet.ServerboundPackets26_1;
 import net.raphimc.viabedrock.ViaBedrock;
-import net.raphimc.viabedrock.api.model.container.Container;
-import net.raphimc.viabedrock.api.model.container.player.InventoryContainer;
-import net.raphimc.viabedrock.api.model.entity.ClientPlayerEntity;
-import net.raphimc.viabedrock.api.model.entity.Entity;
-import net.raphimc.viabedrock.api.util.PacketFactory;
-import net.raphimc.viabedrock.experimental.model.inventory.BedrockInventoryTransaction;
-import net.raphimc.viabedrock.experimental.model.inventory.InventoryActionData;
-import net.raphimc.viabedrock.experimental.model.inventory.InventorySource;
-import net.raphimc.viabedrock.experimental.model.inventory.InventoryTransactionData;
 import net.raphimc.viabedrock.experimental.model.map.MapDecoration;
 import net.raphimc.viabedrock.experimental.model.map.MapObject;
 import net.raphimc.viabedrock.experimental.model.map.MapTrackedObject;
-import net.raphimc.viabedrock.experimental.rewriter.InventoryTransactionRewriter;
 import net.raphimc.viabedrock.experimental.storage.MapTracker;
 import net.raphimc.viabedrock.experimental.util.JavaMapPaletteUtil;
-import net.raphimc.viabedrock.experimental.util.ProtocolUtil;
 import net.raphimc.viabedrock.protocol.BedrockProtocol;
 import net.raphimc.viabedrock.protocol.ClientboundBedrockPackets;
-import net.raphimc.viabedrock.protocol.ServerboundBedrockPackets;
-import net.raphimc.viabedrock.protocol.data.enums.Direction;
 import net.raphimc.viabedrock.protocol.data.enums.bedrock.ClientboundMapItemDataPacket_Type;
-import net.raphimc.viabedrock.protocol.data.enums.bedrock.ComplexInventoryTransaction_Type;
-import net.raphimc.viabedrock.protocol.data.enums.bedrock.ItemUseInventoryTransaction_TriggerType;
-import net.raphimc.viabedrock.protocol.data.enums.bedrock.generated.*;
-import net.raphimc.viabedrock.protocol.data.enums.java.generated.GameMode;
-import net.raphimc.viabedrock.protocol.data.enums.java.generated.InteractionHand;
-import net.raphimc.viabedrock.protocol.data.enums.java.generated.PlayerActionAction;
-import net.raphimc.viabedrock.protocol.model.BedrockItem;
-import net.raphimc.viabedrock.protocol.model.EntityLink;
-import net.raphimc.viabedrock.protocol.model.Position3f;
-import net.raphimc.viabedrock.protocol.storage.ChunkTracker;
-import net.raphimc.viabedrock.protocol.storage.EntityTracker;
-import net.raphimc.viabedrock.protocol.storage.InventoryTracker;
 import net.raphimc.viabedrock.protocol.types.BedrockTypes;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.logging.Level;
 
 /**
  * This class is used to register experimental features that are not yet stable/tested enough to be included in the main protocol.
  * These features may be subject to change or removal in future versions.
+ * <p>
+ * Note: Block breaking/placing, item use and entity metadata/mounting have been promoted to the stable protocol.
+ * What remains experimental is the map item data rewriter and the experimental item data rewriter.
  */
 public class ExperimentalFeatures {
 
     private static final int MAP_FLAGS_ALL = ClientboundMapItemDataPacket_Type.Creation.getValue() | ClientboundMapItemDataPacket_Type.DecorationUpdate.getValue() | ClientboundMapItemDataPacket_Type.TextureUpdate.getValue();
 
     public static void registerPacketTranslators(final BedrockProtocol protocol) {
-        ProtocolUtil.prependServerbound(protocol, ServerboundPackets26_1.PLAYER_ACTION, wrapper -> {
-            final InventoryTransactionRewriter inventoryTransactionRewriter = wrapper.user().get(InventoryTransactionRewriter.class);
-            final InventoryTracker inventoryTracker = wrapper.user().get(InventoryTracker.class);
-
-            final PlayerActionAction action = PlayerActionAction.values()[wrapper.passthrough(Types.VAR_INT)]; // action
-            wrapper.passthrough(Types.BLOCK_POSITION1_14); // block position
-            wrapper.passthrough(Types.UNSIGNED_BYTE); // face
-            wrapper.passthrough(Types.VAR_INT); // sequence number
-
-            if (action == PlayerActionAction.RELEASE_USE_ITEM) {
-                final InventoryContainer inventoryContainer = inventoryTracker.getInventoryContainer();
-
-                wrapper.clearPacket();
-                wrapper.setPacketType(ServerboundBedrockPackets.INVENTORY_TRANSACTION);
-
-                BedrockInventoryTransaction inventoryTransaction = new BedrockInventoryTransaction(
-                        0, // legacy request id
-                        null,
-                        null,
-                        ComplexInventoryTransaction_Type.ItemReleaseTransaction,
-                        new InventoryTransactionData.ReleaseItemTransactionData(
-                                ItemReleaseInventoryTransaction_ActionType.Release,
-                                inventoryContainer.getSelectedHotbarSlot(),
-                                inventoryContainer.getSelectedHotbarItem(),
-                                wrapper.user().get(EntityTracker.class).getClientPlayer().position()
-                        )
-                );
-                wrapper.write(inventoryTransactionRewriter.getInventoryTransactionType(), inventoryTransaction);
-
-                wrapper.sendToServer(BedrockProtocol.class);
-                wrapper.cancel();
-            } else if (action == PlayerActionAction.DROP_ITEM || action == PlayerActionAction.DROP_ALL_ITEMS) {
-                final BedrockItem currentItem = inventoryTracker.getInventoryContainer().getSelectedHotbarItem();
-
-                wrapper.cancel();
-                if (currentItem.isEmpty()) {
-                    return;
-                }
-
-                BedrockItem predictedAmount = currentItem.copy();
-                if (action == PlayerActionAction.DROP_ITEM) {
-                    predictedAmount.setAmount(1); // Drop a single item
-                }
-
-                BedrockItem predictedToItem = currentItem.copy();
-                if (action == PlayerActionAction.DROP_ITEM) {
-                    if (predictedToItem.amount() > 1) {
-                        predictedToItem.setAmount(currentItem.amount() - 1);
-                    } else {
-                        predictedToItem = BedrockItem.empty();
-                    }
-                } else {
-                    predictedToItem = BedrockItem.empty();
-                }
-
-                final PacketWrapper transactionPacket = PacketWrapper.create(ServerboundBedrockPackets.INVENTORY_TRANSACTION, wrapper.user());
-
-                BedrockInventoryTransaction inventoryTransaction = new BedrockInventoryTransaction(
-                        0,
-                        null,
-                        List.of(
-                                new InventoryActionData(
-                                        new InventorySource(InventorySourceType.World_Interaction, ContainerID.CONTAINER_ID_NONE.getValue(), InventorySource_InventorySourceFlags.No_Flag),
-                                        0,
-                                        BedrockItem.empty(),
-                                        predictedAmount
-                                ),
-                                new InventoryActionData(
-                                        new InventorySource(InventorySourceType.Container_Inventory, ContainerID.CONTAINER_ID_INVENTORY.getValue(), InventorySource_InventorySourceFlags.No_Flag),
-                                        inventoryTracker.getInventoryContainer().getSelectedHotbarSlot(),
-                                        currentItem,
-                                        predictedToItem
-                                )
-                        ),
-
-                        ComplexInventoryTransaction_Type.NormalTransaction,
-                        new InventoryTransactionData.NormalTransactionData()
-                );
-
-                transactionPacket.write(inventoryTransactionRewriter.getInventoryTransactionType(), inventoryTransaction);
-
-                transactionPacket.sendToServer(BedrockProtocol.class);
-
-                //TODO: I think vanilla client also sends these and im not sure what their purposes are but it works without them
-                    /*final PacketWrapper interactPacket = PacketWrapper.create(ServerboundBedrockPackets.INTERACT, wrapper.user());
-
-                    interactPacket.write(Types.BYTE, (byte) InteractPacket_Action.InteractUpdate.getValue());
-                    interactPacket.write(BedrockTypes.UNSIGNED_VAR_LONG, clientPlayer.runtimeId());
-                    interactPacket.write(BedrockTypes.POSITION_3F, new Position3f(0, 0, 0));
-
-                    interactPacket.sendToServer(BedrockProtocol.class);
-
-                    final PacketWrapper mobEquipPacket = PacketWrapper.create(ServerboundBedrockPackets.MOB_EQUIPMENT, wrapper.user());
-
-                    mobEquipPacket.write(BedrockTypes.UNSIGNED_VAR_LONG, clientPlayer.runtimeId());
-                    mobEquipPacket.write(itemRewriter.newItemType(), predictedToItem);
-                    mobEquipPacket.write(Types.BYTE, inventoryTracker.getInventoryContainer().getSelectedHotbarSlot());
-                    mobEquipPacket.write(Types.BYTE, inventoryTracker.getInventoryContainer().getSelectedHotbarSlot());
-                    mobEquipPacket.write(Types.BYTE, (byte) ContainerID.CONTAINER_ID_INVENTORY.getValue());
-
-                    mobEquipPacket.sendToServer(BedrockProtocol.class);*/
-            }
-
-
-        });
-
-        // TODO: Track when the player start using item and send the StartUsingItem input data to the server.
-        protocol.registerServerbound(ServerboundPackets26_1.USE_ITEM, ServerboundBedrockPackets.INVENTORY_TRANSACTION, wrapper -> {
-            final EntityTracker entityTracker = wrapper.user().get(EntityTracker.class);
-            final InventoryContainer inventoryContainer = wrapper.user().get(InventoryTracker.class).getInventoryContainer();
-            final InventoryTransactionRewriter inventoryTransactionRewriter = wrapper.user().get(InventoryTransactionRewriter.class);
-
-            final int hand = wrapper.read(Types.VAR_INT); // hand
-            wrapper.read(Types.VAR_INT); // sequence
-            wrapper.read(Types.FLOAT); // yaw
-            wrapper.read(Types.FLOAT); // pitch
-
-            // Bedrock can't hold the majority of item in offhand and can't use any either.
-            // TODO: We need to handle cases where the item changes, or it affect player movement (eg: eating/blocking/etc)
-            if (hand != InteractionHand.MAIN_HAND.ordinal()) {
-                wrapper.cancel();
-                return;
-            }
-
-            BedrockInventoryTransaction inventoryTransaction = new BedrockInventoryTransaction(
-                    0, // legacy request id
-                    null,
-                    null,
-                    ComplexInventoryTransaction_Type.ItemUseTransaction,
-                    new InventoryTransactionData.UseItemTransactionData(
-                            ItemUseInventoryTransaction_ActionType.Use,
-                            ItemUseInventoryTransaction_TriggerType.Unknown,
-                            new BlockPosition(0, 0, 0), // block position
-                            255, // block face
-                            inventoryContainer.getSelectedHotbarSlot(),
-                            inventoryContainer.getSelectedHotbarItem(),
-                            entityTracker.getClientPlayer().position(),
-                            Position3f.ZERO, // click position
-                            0, // block runtime id
-                            ItemUseInventoryTransaction_PredictedResult.Failure,
-                            ItemUseInventoryTransaction_ClientCooldownState.Off
-                    )
-            );
-            wrapper.write(inventoryTransactionRewriter.getInventoryTransactionType(), inventoryTransaction);
-        });
-
-        protocol.registerServerbound(ServerboundPackets26_1.USE_ITEM_ON, null, wrapper -> {
-            wrapper.cancel();
-
-            final ClientPlayerEntity clientPlayer = wrapper.user().get(EntityTracker.class).getClientPlayer();
-            final InventoryTracker inventoryTracker = wrapper.user().get(InventoryTracker.class);
-            final ChunkTracker chunkTracker = wrapper.user().get(ChunkTracker.class);
-            final InventoryTransactionRewriter inventoryTransactionRewriter = wrapper.user().get(InventoryTransactionRewriter.class);
-
-            final InteractionHand hand = InteractionHand.values()[wrapper.read(Types.VAR_INT)]; // hand
-
-            BlockPosition position = wrapper.read(Types.BLOCK_POSITION1_14); // block position
-            int faceInt = wrapper.read(Types.UNSIGNED_BYTE); // face
-            Direction direction = Direction.getFromVerticalId(faceInt);
-            if (direction == null) {
-                ViaBedrock.getPlatform().getLogger().log(Level.WARNING, "Unknown block face id: " + faceInt);
-                return;
-            }
-            BlockFace face = direction.blockFace();
-            Position3f clickPosition = new Position3f(
-                    wrapper.read(Types.FLOAT), // x
-                    wrapper.read(Types.FLOAT), // y
-                    wrapper.read(Types.FLOAT)  // z
-            );
-            boolean insideBlock = wrapper.read(Types.BOOLEAN); // inside block
-            wrapper.read(Types.BOOLEAN); // world border, this doesn't exist on Bedrock.
-
-            // Send back block changed ack with the sequence, this will help with ghost blocks.
-            PacketFactory.sendJavaBlockChangedAck(wrapper.user(), wrapper.read(Types.VAR_INT));
-
-            // The player can only interact using the main hand on Bedrock!
-            if (hand != InteractionHand.MAIN_HAND) {
-                return;
-            }
-
-            // The bedrock client will send a start item use on action to the server first.
-            ExperimentalPacketFactory.sendBedrockPlayerAction(
-                    wrapper.user(),
-                    clientPlayer.runtimeId(),
-                    PlayerActionType.StartItemUseOn,
-                    position,
-                    insideBlock ? position : position.getRelative(face),
-                    faceInt
-            );
-
-            // This is the main packet that the bedrock client use to interact with block.The rest of the
-            final PacketWrapper transactionPacket = PacketWrapper.create(ServerboundBedrockPackets.INVENTORY_TRANSACTION, wrapper.user());
-
-            BedrockItem predictedToItem = inventoryTracker.getInventoryContainer().getSelectedHotbarItem().copy();
-            // This is not entirely correct, but at least it's more accurate than not sending actions or sending the original item data.
-            if (predictedToItem.blockRuntimeId() != 0 && clientPlayer.javaGameMode() != GameMode.CREATIVE) {
-                predictedToItem.setAmount(predictedToItem.amount() - 1);
-            }
-            if (predictedToItem.amount() <= 0) {
-                predictedToItem = BedrockItem.empty();
-            }
-
-            BedrockInventoryTransaction inventoryTransaction = new BedrockInventoryTransaction(
-                    0, // legacy request id
-                    null,
-                    List.of(
-                            new InventoryActionData(
-                                    new InventorySource(InventorySourceType.Container_Inventory, ContainerID.CONTAINER_ID_INVENTORY.getValue(), InventorySource_InventorySourceFlags.No_Flag),
-                                    inventoryTracker.getInventoryContainer().getSelectedHotbarSlot(),
-                                    inventoryTracker.getInventoryContainer().getSelectedHotbarItem(),
-                                    predictedToItem
-                            )
-                    ),
-                    ComplexInventoryTransaction_Type.ItemUseTransaction,
-                    new InventoryTransactionData.UseItemTransactionData(
-                            ItemUseInventoryTransaction_ActionType.Place,
-                            ItemUseInventoryTransaction_TriggerType.PlayerInput,
-                            position,
-                            faceInt,
-                            inventoryTracker.getInventoryContainer().getSelectedHotbarSlot(),
-                            inventoryTracker.getInventoryContainer().getSelectedHotbarItem(),
-                            clientPlayer.position(),
-                            clickPosition,
-                            chunkTracker.getBlockState(position),
-                            ItemUseInventoryTransaction_PredictedResult.Success,
-                            ItemUseInventoryTransaction_ClientCooldownState.Off
-                    )
-            );
-            transactionPacket.write(inventoryTransactionRewriter.getInventoryTransactionType(), inventoryTransaction);
-
-            transactionPacket.sendToServer(BedrockProtocol.class);
-
-            // Bedrock sends a stop item use on after the transaction packet
-            ExperimentalPacketFactory.sendBedrockPlayerAction(
-                    wrapper.user(),
-                    clientPlayer.runtimeId(),
-                    PlayerActionType.StopItemUseOn,
-                    position,
-                    new BlockPosition(0, 0, 0),
-                    0
-            );
-        });
-        protocol.registerClientbound(ClientboundBedrockPackets.INVENTORY_TRANSACTION, null, wrapper -> {
-            final InventoryTransactionRewriter inventoryTransactionRewriter = wrapper.user().get(InventoryTransactionRewriter.class);
-            InventoryTracker inventoryTracker = wrapper.user().get(InventoryTracker.class);
-
-            wrapper.cancel();
-            BedrockInventoryTransaction inventoryTransaction = wrapper.read(inventoryTransactionRewriter.getInventoryTransactionType());
-
-            if (inventoryTransaction.legacyRequestId() != 0) {
-                // Ignore legacy inventory transactions for now
-                return;
-            }
-
-            if (inventoryTransaction.actions() != null && !inventoryTransaction.actions().isEmpty()) {
-                for (InventoryActionData action : inventoryTransaction.actions()) {
-                    if (action.source().type() == InventorySourceType.Container_Inventory) {
-                        Container container = inventoryTracker.getContainerClientbound((byte) action.source().containerId(), null, null);
-
-                        if (container != null) {
-                            container.setItem(action.slot(), action.toItem());
-                            PacketFactory.sendJavaContainerSetContent(wrapper.user(),  container);
-                        } else {
-                            ViaBedrock.getPlatform().getLogger().log(Level.WARNING, "Received inventory action for unknown container ID: " + action.source().containerId());
-                        }
-                    }
-                }
-            }
-
-            switch (inventoryTransaction.transactionType()) {
-                case NormalTransaction -> {
-                    break; // Nothing to do here for now
-                }
-                default -> {
-                    ViaBedrock.getPlatform().getLogger().log(Level.WARNING, "Received unsupported inventory transaction type: " + inventoryTransaction.transactionType());
-                }
-            }
-        });
-        // SET_ENTITY_LINK handling has been promoted to the stable path (EntityPackets)
-
         protocol.registerClientbound(ClientboundBedrockPackets.MAP_ITEM_DATA, ClientboundPackets26_1.MAP_ITEM_DATA, wrapper -> {
             wrapper.cancel();
             wrapper.clearPacket();
@@ -546,7 +241,6 @@ public class ExperimentalFeatures {
     }
 
     public static void registerStorages(final UserConnection user) {
-        user.put(new InventoryTransactionRewriter(user));
         user.put(new MapTracker(user));
     }
 }
