+++ src/main/java/net/raphimc/viabedrock/protocol/BedrockProtocol.java
@@ -47,6 +47,7 @@ import net.raphimc.viabedrock.protocol.provider.BlobCacheProvider;
 import net.raphimc.viabedrock.protocol.provider.NettyPipelineProvider;
 import net.raphimc.viabedrock.protocol.provider.ResourcePackProvider;
 import net.raphimc.viabedrock.protocol.provider.SkinProvider;
+import net.raphimc.viabedrock.protocol.rewriter.InventoryTransactionRewriter;
 import net.raphimc.viabedrock.protocol.storage.*;
 import net.raphimc.viabedrock.protocol.task.*;
 
@@ -155,6 +156,7 @@ public class BedrockProtocol extends StatelessTransitionProtocol<ClientboundBedr
         user.put(new ScoreboardTracker());
         user.put(new InventoryTracker(user));
         user.put(new BreakingTracker(user));
+        user.put(new InventoryTransactionRewriter(user));
 
         if (ViaBedrock.getConfig().shouldEnableExperimentalFeatures()) {
             ExperimentalFeatures.registerStorages(user);
